feat: 상품이미지 - 유효한 회원 정보 연결

+++ src/main/java/com/team01/billage/product/service/ProductImageService.java
@@ -1,10 +1,13 @@
 package com.team01.billage.product.service;
 
+import com.team01.billage.exception.CustomException;
 import com.team01.billage.product.domain.Product;
 import com.team01.billage.product.domain.ProductImage;
 import com.team01.billage.product.dto.ProductImageDeleteRequestDto;
 import com.team01.billage.product.dto.ProductImageRequestDto;
 import com.team01.billage.product.repository.ProductImageRepository;
+import com.team01.billage.product.repository.ProductRepository;
+import com.team01.billage.user.domain.Users;
 import com.team01.billage.utils.s3.S3BucketService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,11 +16,14 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static com.team01.billage.exception.ErrorCode.*;
+
 @Service
 @RequiredArgsConstructor
 public class ProductImageService {
 
     private final ProductImageRepository productImageRepository;
+    private final ProductRepository productRepository;
     private final S3BucketService s3BucketService;
 
     // 상품 이미지 저장
@@ -47,7 +53,18 @@ public class ProductImageService {
 
     // 특정 이미지 삭제
     @Transactional
-    public void deleteProductImages(List<ProductImageDeleteRequestDto> productImageDeleteRequestDtos){
+    public void deleteProductImages(
+            Users user,
+            String productId,
+            List<ProductImageDeleteRequestDto> productImageDeleteRequestDtos){
+
+        Product product = productRepository.findById(Long.parseLong(productId))
+                .orElseThrow(() -> new CustomException(PRODUCT_NOT_FOUND));
+
+        if(!product.getSeller().getId().equals(user.getId())){
+            throw new CustomException(ACCESS_DENIED);
+        }
+
         // s3에서 삭제
         for (ProductImageDeleteRequestDto image : productImageDeleteRequestDtos) {
             s3BucketService.delete(image.getImageUrl());
